+++ resources/views/admin/clientes/index.blade.php
@@ -34,7 +34,7 @@
       <tbody>
       @foreach($clientes as $cliente)
         <tr>
-            <th><center>{{ ++$i }}</center></th>
+            <th><center>M{{str_pad($cliente->id, 6, "0", STR_PAD_LEFT)}}</center></th>
             <td>{{$cliente->cliente_nome }}</td>
             <td>{{$cliente->client_country_city->country_name}}</td>
             <td>{{$cliente->client_country_city->state_name}}</td>
@@ -53,6 +53,7 @@
                  <a href="{{ route('clientes.edit', $cliente->id)}}" class="btn btn-primary btn-xs"><i class="fa fa-fw fa-pencil"></i></a>
                  <a href="{{ route ('clientes.show', $cliente->id)}}" class="btn btn-warning btn-xs"><i class="fa fa-fw fa-info-circle"></i></a>
                  <a href="{{ url ('admin/financas/extrato', $cliente->token_id)}}" class="btn btn-success btn-xs"><i class="fa fa-pie-chart"></i></a>
+                 @can('clientes_destroy')
                   @if($cliente->status =='1')
                   {!! Form::open(['method' => 'DELETE','route' => ['clientes.destroy', $cliente->id],'style'=>'display:inline']) !!}
                   {!! Form::button('<i class="fa fa-trash-o"></i>', ['class'=>'btn btn-danger btn-xs', 'type'=>'submit']) !!}
@@ -63,6 +64,7 @@
                   {!! Form::button('<i class="fa fa-recycle"></i>', ['class'=>'btn btn-danger btn-xs', 'type'=>'submit']) !!}
                   {!! Form::close() !!}
                   @endif
+                 @endcan
              </center></td>
              <td><center> <a href="{{url('admin/contrato',$cliente->id)}}" class="btn btn-danger btn-xs">Contrato</a>
                  <a href="{{url('admin/prospecoes',$cliente->id)}}" class="btn btn-success  btn-xs">Prospeção</a>
